add support for configurable bpm
- pass bpm as u32 for easiser conversion
- add utils for converting bpm to tempo bytes
- update song midify methods

+++ src/lib.rs
@@ -26,7 +26,7 @@ pub struct Config {
 }
 
 impl Config {
-    pub fn new(tempo: u16, out_file: String) -> Config {
+    pub fn new(tempo: u32, out_file: String) -> Config {
         Config {
             song_config: SongConfig { 
                 tempo,
